requirements.txt
python-dotenv==1.0.1
websocket-client==1.8.0